backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# Celery & Redis (optional — only needed for scheduled pipeline)
celery==5.4.0
redis==5.1.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9   # PostgreSQL driver
pymysql==1.1.1             # MySQL driver
alembic==1.13.2

# Data Ingestion
telethon==1.37.0
youtube-comment-downloader==0.1.78
tweepy==4.14.0
praw==7.7.1
yfinance>=1.1.0

# NLP & ML
transformers==4.45.0
torch==2.4.0
peft==0.13.0
datasets==3.0.0
scikit-learn==1.5.2
nltk==3.9.1
spacy==3.8.2
emoji==2.12.1

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Auth
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12

# Utilities
google-generativeai==0.8.3
loguru==0.7.2
